Init docker-compose config file
A bit of this is hard coded and messy, but as I get more familiar with
docker compose, I plan to clean it up along the way. I'm not going to
worry about deploying to ECS right now. I realized I've spent too long
attempting to master it, but it's not entirely necessary at this stage.